LPS-53627 extract related assets functionality to its own module
