+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-related-assets-asset-addon-entry/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay Journal Content Related Assets Asset Addon Entry
+Bundle-SymbolicName: com.liferay.journal.content.asset.addon.entry.related.assets
+Bundle-Version: 1.0.0
+Include-Resource:\
+	classes,\
+	META-INF/resources=src/META-INF/resources
+Web-ContextPath: /journal-content-related-assets-asset-addon-entry
